+++ src/propedit/options.cpp
@@ -51,19 +51,6 @@ void
 options_c::prune_empty_masters() {
   std::unordered_map<EbmlMaster *, bool> handled;
 
-  auto handler = [&handled](EbmlMaster *parent, EbmlMaster *child) {
-    if (!parent || !child || handled[child] || (0 < child->ListSize()))
-      return;
-
-    handled[child] = true;
-
-    auto itr = std::find(parent->begin(), parent->end(), child);
-    if (itr != parent->end())
-      parent->Remove(itr);
-
-    delete child;
-  };
-
   for (auto &target : m_targets) {
     if (!dynamic_cast<track_target_c *>(target.get()))
       continue;
@@ -72,11 +59,11 @@ options_c::prune_empty_masters() {
     auto masters       = track_target.get_masters();
 
     for (auto const &change : track_target.m_changes) {
-      handler(change->m_sub_sub_master, change->m_sub_sub_sub_master); // sub_sub_master and sub_sub_sub_master
-      handler(std::get<1>(masters),     change->m_sub_sub_master);     // sub_master     and sub_sub_master
+      remove_master_from_parent_if_empty_or_only_defaults(change->m_sub_sub_master, change->m_sub_sub_sub_master, handled); // sub_sub_master and sub_sub_sub_master
+      remove_master_from_parent_if_empty_or_only_defaults(std::get<1>(masters),     change->m_sub_sub_master,     handled); // sub_master     and sub_sub_master
     }
 
-    handler(std::get<0>(masters), std::get<1>(masters));               // master         and sub_master
+    remove_master_from_parent_if_empty_or_only_defaults(std::get<0>(masters), std::get<1>(masters), handled);               // master         and sub_master
   }
 }